requirements.txt
requests
qrcode
speedtest-cli
